requirements.txt
django==4.1.6
dj-rest-auth
django-rest-framework